Use more OOP to abstract certain functions and classes out
